+++ src/ContaoManager/Plugin.php
@@ -12,14 +12,27 @@ use Contao\CoreBundle\ContaoCoreBundle;
 use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
+use Contao\ManagerPlugin\Config\ConfigPluginInterface;
 use HeimrichHannot\RequestBundle\HeimrichHannotContaoRequestBundle;
+use Symfony\Component\Config\Loader\LoaderInterface;
 
-class Plugin implements BundlePluginInterface
+class Plugin implements BundlePluginInterface, ConfigPluginInterface
 {
     public function getBundles(ParserInterface $parser)
     {
         return [
             BundleConfig::create(HeimrichHannotContaoRequestBundle::class)->setLoadAfter([ContaoCoreBundle::class]),
         ];
+    }
+
+    /**
+     * Allows a plugin to load container configuration.
+     * @param LoaderInterface $loader
+     * @param array $managerConfig
+     * @throws \Exception
+     */
+    public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig)
+    {
+        $loader->load('@HeimrichHannotContaoRequestBundle/Ressources/config/services.yml');
     }
 }